pyproject.toml
[project]
name = "document-q-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "huggingface-hub>=0.34.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "scikit-learn>=1.7.1",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "transformers>=4.55.3",
]
